[step2] 패키지 분리
