+++ src/main/kotlin/calculator/Calculator.kt
@@ -1,16 +1,4 @@
-inline fun Sequence<MatchResult>.reduce(operation: (acc: Int, operator: String, value: Int) -> Int): Int {
-    var accumulator = 0
-
-    for (element in this) {
-        accumulator = if (element.groupValues[1].isNotEmpty())
-            element.groupValues[1].toInt()
-        else accumulator
-
-        accumulator = operation(accumulator, element.groupValues[2], element.groupValues[3].toInt())
-    }
-    return accumulator
-}
-
+package calculator
 
 object Calculator {
     private val EXPRESSION_REGEX = """(\d*) ([+\-*/]) (\d*)""".toRegex()
