+++ src/test/kotlin/calculator/CalculatorTest.kt
@@ -1,3 +1,5 @@
+package calculator
+
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
